layers: Rename BothRangeMap
