+++ BUILD.gn
@@ -106,6 +106,7 @@ vvl_sources = [
   "layers/containers/container_utils.h",
   "layers/containers/custom_containers.h",
   "layers/containers/limits.h",
+  "layers/containers/array_range_map.h",
   "layers/containers/small_container.h",
   "layers/containers/small_range_map.h",
   "layers/containers/small_vector.h",
