+++ layers/state_tracker/subresource_adapter.h
@@ -24,8 +24,6 @@
 #include <variant>
 #include <vector>
 #include "containers/range.h"
-#include "containers/range_map.h"
-#include "containers/small_range_map.h"
 #include "containers/small_vector.h"
 #include "vulkan/vulkan.h"
 
@@ -138,11 +136,11 @@ class RangeEncoder {
     inline IndexType SubresourceCount() const { return AspectSize() * Limits().aspect_index; }
     inline VkImageAspectFlags AspectMask() const { return limits_.aspectMask; }
     inline VkImageAspectFlagBits AspectBit(uint32_t aspect_index) const {
-        RANGE_ASSERT(aspect_index < limits_.aspect_index);
+        assert(aspect_index < limits_.aspect_index);
         return aspect_bits_[aspect_index];
     }
     inline IndexType AspectBase(uint32_t aspect_index) const {
-        RANGE_ASSERT(aspect_index < limits_.aspect_index);
+        assert(aspect_index < limits_.aspect_index);
         return aspect_base_[aspect_index];
     }
 
@@ -461,240 +459,6 @@ class ImageRangeGenerator {
     IncrementerState incr_state_;
     bool single_full_size_range_ = true;
     bool is_depth_sliced_ = false;
-};
-
-// double wrapped map variants.. to avoid needing to templatize on the range map type.  The underlying maps are available for
-// use in performance sensitive places that are *already* templatized (for example update_range_value).
-// In STL style.  Note that N must be < uint8_t max
-template <typename T, size_t N>
-class BothRangeMap {
-    using SmallMap = sparse_container::small_range_map<IndexType, T, vvl::range<IndexType>, N>;
-    using SmallMapIterator = typename SmallMap::iterator;
-    using SmallMapConstIterator = typename SmallMap::const_iterator;
-
-    using BigMap = sparse_container::range_map<IndexType, T>;
-    using BigMapIterator = typename BigMap::iterator;
-    using BigMapConstIterator = typename BigMap::const_iterator;
-
-  public:
-    using index_type = IndexType;
-    using key_type = vvl::range<IndexType>;
-    using mapped_type = T;
-    using value_type = std::pair<const key_type, mapped_type>;
-
-    template <typename Value, typename SmallIt, typename BigIt>
-    class IteratorImpl {
-      public:
-        Value* operator->() const {
-            if (is_small_it_) {
-                return small_it_.operator->();
-            } else {
-                return big_it_.operator->();
-            }
-        }
-        Value& operator*() const {
-            if (is_small_it_) {
-                return small_it_.operator*();
-            } else {
-                return big_it_.operator*();
-            }
-        }
-        IteratorImpl& operator++() {
-            if (is_small_it_) {
-                small_it_.operator++();
-            } else {
-                big_it_.operator++();
-            }
-            return *this;
-        }
-        IteratorImpl& operator--() {
-            if (is_small_it_) {
-                small_it_.operator--();
-            } else {
-                big_it_.operator--();
-            }
-            return *this;
-        }
-        IteratorImpl& operator=(const IteratorImpl& other) {
-            is_small_it_ = other.is_small_it_;
-            small_it_ = other.small_it_;
-            big_it_ = other.big_it_;
-            return *this;
-        }
-        bool operator==(const IteratorImpl& other) const {
-            // It's enough just to compare both iterators.
-            return small_it_ == other.small_it_ && big_it_ == other.big_it_;
-        }
-        bool operator!=(const IteratorImpl& other) const { return !(*this == other); }
-
-        IteratorImpl() = default;
-        IteratorImpl(const IteratorImpl& other) = default;
-        IteratorImpl(const SmallIt& it) : is_small_it_(true), small_it_(it) {}
-        IteratorImpl(const BigIt& it) : is_small_it_(false), big_it_(it) {}
-
-      private:
-        friend BothRangeMap;
-
-        bool is_small_it_ = false;
-        SmallIt small_it_;
-        BigIt big_it_;
-    };
-
-    using iterator = IteratorImpl<value_type, SmallMapIterator, BigMapIterator>;
-    // TODO change const iterator to derived class if iterator -> const_iterator constructor is needed
-    using const_iterator = IteratorImpl<const value_type, SmallMapConstIterator, BigMapConstIterator>;
-
-    iterator begin() {
-        if (UsesSmallMap()) {
-            return iterator(GetSmallMap().begin());
-        } else {
-            return iterator(GetBigMap().begin());
-        }
-    }
-    const_iterator cbegin() const {
-        if (UsesSmallMap()) {
-            return const_iterator(GetSmallMap().begin());
-        } else {
-            return const_iterator(GetBigMap().begin());
-        }
-    }
-    const_iterator begin() const { return cbegin(); }
-
-    iterator end() {
-        if (UsesSmallMap()) {
-            return iterator(GetSmallMap().end());
-        } else {
-            return iterator(GetBigMap().end());
-        }
-    }
-    const_iterator cend() const {
-        if (UsesSmallMap()) {
-            return const_iterator(GetSmallMap().end());
-        } else {
-            return const_iterator(GetBigMap().end());
-        }
-    }
-    const_iterator end() const { return cend(); }
-
-    iterator find(const key_type& key) {
-        if (UsesSmallMap()) {
-            return iterator(GetSmallMap().find(key));
-        } else {
-            return iterator(GetBigMap().find(key));
-        }
-    }
-
-    const_iterator find(const key_type& key) const {
-        if (UsesSmallMap()) {
-            return const_iterator(GetSmallMap().find(key));
-        } else {
-            return const_iterator(GetBigMap().find(key));
-        }
-    }
-
-    iterator find(const index_type& index) {
-        if (UsesSmallMap()) {
-            return iterator(GetSmallMap().find(index));
-        } else {
-            return iterator(GetBigMap().find(index));
-        }
-    }
-
-    const_iterator find(const index_type& index) const {
-        if (UsesSmallMap()) {
-            return const_iterator(GetSmallMap().find(index));
-        } else {
-            return const_iterator(GetBigMap().find(index));
-        }
-    }
-
-    // TODO -- this is supposed to be a const_iterator, which is constructable from an iterator
-    void insert(const iterator& hint, const value_type& value) {
-        if (UsesSmallMap()) {
-            assert(hint.is_small_it_);
-            GetSmallMap().insert(hint.small_it_, value);
-        } else {
-            assert(!hint.is_small_it_);
-            GetBigMap().insert(hint.big_it_, value);
-        }
-    }
-
-    iterator lower_bound(const key_type& key) {
-        if (UsesSmallMap()) {
-            return iterator(GetSmallMap().lower_bound(key));
-        } else {
-            return iterator(GetBigMap().lower_bound(key));
-        }
-    }
-
-    const_iterator lower_bound(const key_type& key) const {
-        if (UsesSmallMap()) {
-            return const_iterator(GetSmallMap().lower_bound(key));
-        } else {
-            return const_iterator(GetBigMap().lower_bound(key));
-        }
-    }
-
-    template <typename Value>
-    iterator overwrite_range(const iterator& lower, Value&& value) {
-        if (UsesSmallMap()) {
-            assert(lower.is_small_it_);
-            return GetSmallMap().overwrite_range(lower.small_it_, std::forward<Value>(value));
-        } else {
-            assert(!lower.is_small_it_);
-            return GetBigMap().overwrite_range(lower.big_it_, std::forward<Value>(value));
-        }
-    }
-
-    // With power comes responsibility (🕷).  You can get to the underlying maps, s.t. in inner loops, the "SmallMode" checks can be
-    // avoided per call, just be sure and Get the correct one.
-    const SmallMap& GetSmallMap() const {
-        assert(UsesSmallMap());
-        return std::get<SmallMap>(map_);
-    }
-    SmallMap& GetSmallMap() {
-        assert(UsesSmallMap());
-        return std::get<SmallMap>(map_);
-    }
-    const BigMap& GetBigMap() const {
-        assert(!UsesSmallMap());
-        return std::get<BigMap>(map_);
-    }
-    BigMap& GetBigMap() {
-        assert(!UsesSmallMap());
-        return std::get<BigMap>(map_);
-    }
-
-    BothRangeMap() = delete;
-
-    BothRangeMap(index_type limit) {
-        if (limit <= N) {
-            map_ = SmallMap(limit);
-        } else {
-            map_ = BigMap();
-        }
-    }
-
-    bool empty() const {
-        if (UsesSmallMap()) {
-            return GetSmallMap().empty();
-        } else {
-            return GetBigMap().empty();
-        }
-    }
-
-    size_t size() const {
-        if (UsesSmallMap()) {
-            return GetSmallMap().size();
-        } else {
-            return GetBigMap().size();
-        }
-    }
-
-    bool UsesSmallMap() const { return std::holds_alternative<SmallMap>(map_); }
-
-  private:
-    std::variant<SmallMap, BigMap> map_;
 };
 
 }  // namespace subresource_adapter
